+++ packages/theme/plugin.js
@@ -4,7 +4,6 @@ K.Plugin({
 	templates: {
 		pageHome: 'pageHome_theme',
 		footer: {
-			'show': false,
 			'footer_theme':{ order: 10 }
 		}
 	},