+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
-...
+- panelList support array in field sortBy
+- fix panelUser template
 
 ## v1.5.2
 - new settings public.map.checkinAutomatic