+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 ## Next version
 - panelList support array in field sortBy
 - fix panelUser template
+- block email send if user created by admin
 
 ## v1.5.2
 - new settings public.map.checkinAutomatic