+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- new settings to disable user markers settings.public.map.layerUsers.enabled
 - disable placeholders by settings.json
 - fix panelUser template
 - block email send if user created by admin